Refactor EditForm into Own Component for Separation of Concerns. Trying to clean up all the junk code that AI created that isn't used anywhere and simplify the rrule stuff.
