+++ src/app/(tabs)/schedule/[id]/edit.tsx
@@ -1,12 +1,8 @@
 import {
   useDeleteTalentBlockout,
   useGetTalentBlockoutById,
-  useUpdateTalentBlockout,
 } from "@/api/blockouts_api";
-import ScheduleDateTimeCard from "@/components/schedule/ScheduleDateTimeCard";
-import ScheduleDescriptionCard from "@/components/schedule/ScheduleDescriptionCard";
-import ScheduleRecurringCard from "@/components/schedule/ScheduleRecurringCard";
-import ScheduleTitleCard from "@/components/schedule/ScheduleTitleCard";
+import EditScheduleForm from "@/components/schedule/EditScheduleForm";
 import {
   Button,
   ButtonIcon,
@@ -16,33 +12,17 @@ import {
 import { Text } from "@/components/ui/text";
 import { VStack } from "@/components/ui/vstack";
 import { canEditBlockout } from "@/utils/blockout-permissions";
-import { getDayjsFromUtcDateString } from "@/utils/date";
-import {
-  createRRuleFromOptions,
-  getRRuleOptions,
-  updateBlockoutSchema,
-} from "@/validators/blockouts.validators";
-import { zodResolver } from "@hookform/resolvers/zod";
-import dayjs from "dayjs";
-import timezone from "dayjs/plugin/timezone";
-import utc from "dayjs/plugin/utc";
 import { Stack, useLocalSearchParams, useRouter } from "expo-router";
 import { Trash2 } from "lucide-react-native";
-import { useEffect } from "react";
-import { useForm } from "react-hook-form";
-import { Alert, ScrollView } from "react-native";
+import { Alert } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { RRule } from "rrule";
-
-// Enable timezone plugins
-dayjs.extend(utc);
-dayjs.extend(timezone);
 
 export default function EditBlockoutScreen() {
   const router = useRouter();
   const params = useLocalSearchParams<{
     id: string;
   }>();
+  const { mutateAsync: deleteBlockout, isPending } = useDeleteTalentBlockout();
 
   const blockoutId = params.id ? parseInt(params.id) : 0;
 
@@ -51,56 +31,9 @@ export default function EditBlockoutScreen() {
     isLoading,
     error,
   } = useGetTalentBlockoutById(blockoutId, !!blockoutId);
-  const { mutateAsync: updateBlockout } = useUpdateTalentBlockout();
-  const { mutateAsync: deleteBlockout, isPending } = useDeleteTalentBlockout();
 
-  const {
-    control,
-    handleSubmit,
-    setValue,
-    watch,
-    reset,
-    formState: { isSubmitting, errors },
-  } = useForm({
-    resolver: zodResolver(updateBlockoutSchema),
-    defaultValues: {
-      title: "",
-      description: "",
-      start_time: dayjs.utc().startOf("hour").add(1, "hour").toISOString(),
-      end_time: dayjs.utc().startOf("hour").add(2, "hour").toISOString(),
-      timezone: dayjs.tz.guess(),
-      is_all_day: false,
-      rrule: undefined,
-    },
-  });
-
-  // Update form when blockout data is loaded
-  useEffect(() => {
-    if (blockout) {
-      // Convert rrule string from database to RRuleOptions
-      let rruleOptions = null;
-      if (blockout.rrule) {
-        try {
-          const rule = RRule.fromString(blockout.rrule);
-          rruleOptions = getRRuleOptions(rule);
-        } catch (error) {
-          console.warn("Failed to parse RRULE:", error);
-        }
-      }
-
-      reset({
-        title: blockout.title,
-        description: blockout.description || "",
-        start_time: getDayjsFromUtcDateString(
-          blockout.start_time
-        ).toISOString(),
-        end_time: getDayjsFromUtcDateString(blockout.end_time).toISOString(),
-        timezone: blockout.timezone,
-        is_all_day: blockout.is_all_day,
-        rrule: rruleOptions,
-      });
-    }
-  }, [blockout, reset]);
+  // Check if blockout can be edited (end time must be in the future)
+  const canEdit = blockout ? canEditBlockout(blockout) : false;
 
   const deleteButton = () => (
     <Button
@@ -133,15 +66,6 @@ export default function EditBlockoutScreen() {
     </Button>
   );
 
-  // Check if blockout can be edited (end time must be in the future)
-  const canEdit = blockout ? canEditBlockout(blockout) : false;
-
-  const isAllDay = watch("is_all_day");
-  const startTime = watch("start_time");
-  const endTime = watch("end_time");
-  const currentRRule = watch("rrule");
-
-  // Handle loading and error states
   if (isLoading) {
     return (
       <SafeAreaView edges={["bottom"]} className="flex-1 bg-background-0">
@@ -204,39 +128,6 @@ export default function EditBlockoutScreen() {
     );
   }
 
-  const onSubmit = async (data: any) => {
-    try {
-      if (!blockoutId) {
-        Alert.alert("Error", "Blockout ID is missing");
-        return;
-      }
-
-      // Convert RRuleOptions to string for API
-      let rruleString: string | null = null;
-      if (data.rrule) {
-        const rule = createRRuleFromOptions(data.rrule);
-        rruleString = rule.toString(); // Make sure we get the DTSTART and the RRULE part
-      }
-
-      await updateBlockout({
-        blockoutId,
-        updates: {
-          title: data.title,
-          description: data.description || undefined,
-          start_time: data.start_time,
-          end_time: data.end_time,
-          timezone: data.timezone,
-          is_all_day: data.is_all_day || false,
-          rrule: rruleString,
-        },
-      });
-
-      router.back();
-    } catch {
-      Alert.alert("Error", "Failed to update blockout");
-    }
-  };
-
   return (
     <>
       <SafeAreaView edges={["bottom"]} className="flex-1 bg-background-0">
@@ -246,65 +137,7 @@ export default function EditBlockoutScreen() {
             headerRight: canEdit ? deleteButton : undefined,
           }}
         />
-
-        <ScrollView className="flex-1" showsVerticalScrollIndicator={false}>
-          <VStack space="lg" className="p-6">
-            {/* Header Card */}
-            <VStack
-              space="sm"
-              className="bg-white rounded-2xl p-6 border border-primary-200 shadow-sm"
-            >
-              <Text size="xl" bold className="text-typography-900 mb-2">
-                Edit Blockout
-              </Text>
-              <Text className="text-typography-600">
-                Update your blockout details
-              </Text>
-            </VStack>
-
-            {/* Title Card */}
-            <ScheduleTitleCard control={control} />
-
-            {/* Description Card */}
-            <ScheduleDescriptionCard control={control} />
-
-            {/* All Day & Date/Time Card */}
-            <ScheduleDateTimeCard
-              control={control}
-              setValue={setValue}
-              isAllDay={isAllDay ?? false}
-            />
-
-            {/* Recurring Card */}
-            <ScheduleRecurringCard
-              control={control}
-              setValue={setValue}
-              startTime={startTime || dayjs.utc().toISOString()}
-              endTime={endTime || dayjs.utc().toISOString()}
-              currentRRule={currentRRule || null}
-              errors={errors}
-            />
-
-            {/* Submit Button Card */}
-            <VStack className="bg-white rounded-2xl p-6 border border-outline-200 shadow-sm">
-              <Button
-                size="lg"
-                variant="solid"
-                action="primary"
-                onPress={handleSubmit(onSubmit)}
-                isDisabled={isSubmitting}
-                className="rounded-xl"
-              >
-                <ButtonText className="font-semibold">
-                  {isSubmitting ? "Updating..." : "Update Blockout"}
-                </ButtonText>
-              </Button>
-            </VStack>
-          </VStack>
-
-          {/* Bottom padding */}
-          <VStack className="h-8" />
-        </ScrollView>
+        <EditScheduleForm blockoutData={blockout} />
       </SafeAreaView>
     </>
   );
