+++ src/validators/blockouts.validators.ts
@@ -28,9 +28,9 @@ const rruleOptionsSchema = z
     ]),
     dtstart: z.date(),
     interval: z.number().int().positive(),
-    byweekday: z.union([z.number(), z.array(z.number())]).optional(),
-    bymonthday: z.union([z.number(), z.array(z.number())]).optional(),
-    bymonth: z.union([z.number(), z.array(z.number())]).optional(),
+    byweekday: z.array(z.number()).optional(),
+    bymonthday: z.array(z.number()).optional(),
+    bymonth: z.array(z.number()).optional(),
     count: z.number().int().positive().optional(),
     until: z.date().optional(),
   })
@@ -93,6 +93,7 @@ const baseBlockoutSchema = z.object({
   description: z
     .string()
     .max(500, "Description must be less than 500 characters")
+    .nullable()
     .optional(),
 });
 
@@ -136,69 +137,11 @@ export const updateBlockoutSchema = baseBlockoutSchema
     }
   );
 
-// Schema for date range queries
-export const dateRangeSchema = z
-  .object({
-    startDate: z.string().datetime("Invalid start date format"),
-    endDate: z.string().datetime("Invalid end date format"),
-  })
-  .refine((data) => new Date(data.startDate) <= new Date(data.endDate), {
-    message: "End date must be after or equal to start date",
-    path: ["endDate"],
-  });
-
-// Schema for availability check
-export const availabilityCheckSchema = z
-  .object({
-    talentId: z.string().uuid("Invalid talent ID"),
-    startTime: z.string().datetime("Invalid start time format"),
-    endTime: z.string().datetime("Invalid end time format"),
-  })
-  .refine((data) => new Date(data.startTime) < new Date(data.endTime), {
-    message: "End time must be after start time",
-    path: ["endTime"],
-  });
-
-// Export RRULE validation schema
-export const rruleSchema = rruleOptionsSchema;
-
-// Schema for recurring pattern creation helper
-export const recurrencePatternSchema = z
-  .object({
-    frequency: z.enum(["DAILY", "WEEKLY", "MONTHLY", "YEARLY"]),
-    interval: z.number().min(1).max(999).optional(),
-    byWeekDay: z.array(z.number().min(0).max(6)).optional(),
-    until: z.string().datetime().optional(),
-    count: z.number().min(1).max(1000).optional(),
-  })
-  .refine(
-    (data) => {
-      // Either until or count, not both
-      return !(data.until && data.count);
-    },
-    {
-      message: "Cannot specify both 'until' and 'count' for recurrence",
-    }
-  );
-
 // Types inferred from schemas
 export type RRuleOptions = z.infer<typeof rruleOptionsSchema>;
 export type CreateBlockoutInput = z.infer<typeof createBlockoutSchema>;
 export type CreateBlockoutSchemaType = z.infer<typeof createBlockoutSchema>;
 export type UpdateBlockoutInput = z.infer<typeof updateBlockoutSchema>;
-export type DateRangeInput = z.infer<typeof dateRangeSchema>;
-export type AvailabilityCheckInput = z.infer<typeof availabilityCheckSchema>;
-export type RecurrencePatternInput = z.infer<typeof recurrencePatternSchema>;
-
-// Helper function to validate RRULE options
-export const validateRRule = (rrule: RRuleOptions): boolean => {
-  try {
-    rruleSchema.parse(rrule);
-    return true;
-  } catch {
-    return false;
-  }
-};
 
 // Helper function to convert RRuleOptions to RRule instance
 export const createRRuleFromOptions = (options: RRuleOptions): RRule => {
@@ -212,39 +155,10 @@ export const getRRuleOptions = (rule: RRule): RRuleOptions => {
     freq: opts.freq,
     dtstart: opts.dtstart!,
     interval: opts.interval ?? 1,
-    byweekday: opts.byweekday as number | number[] | undefined,
-    bymonthday: opts.bymonthday as number | number[] | undefined,
-    bymonth: opts.bymonth as number | number[] | undefined,
+    byweekday: opts.byweekday as number[] | undefined,
+    bymonthday: opts.bymonthday as number[] | undefined,
+    bymonth: opts.bymonth as number[] | undefined,
     count: opts.count ?? undefined,
     until: opts.until ?? undefined,
-  };
-};
-
-// Helper function to create form-friendly default values
-export const getDefaultBlockoutValues = (): Partial<CreateBlockoutInput> => ({
-  title: "",
-  description: "",
-  is_all_day: false,
-});
-
-// Helper function to format dates for form inputs
-export const formatDateForInput = (date: Date): string => {
-  return date.toISOString().slice(0, 16); // YYYY-MM-DDTHH:MM format for datetime-local input
-};
-
-// Helper function to format dates for all-day events
-export const formatAllDayDate = (
-  date: Date,
-  timezone: string = "UTC"
-): { start_time: string; end_time: string } => {
-  const start = new Date(date);
-  start.setHours(0, 0, 0, 0);
-
-  const end = new Date(date);
-  end.setHours(23, 59, 59, 999);
-
-  return {
-    start_time: start.toISOString(),
-    end_time: end.toISOString(),
   };
 };
